feat(logout):solved tests
tests are running 100%
feature is in finished
