--- src/tests/app.testi.js
@@ -1,21 +1,0 @@
-/** @format */
-
-import app from "../app";
-import chai from "chai";
-import chaiHttp from "chai-http";
-
-chai.use(chaiHttp);
-const expect = require("chai").expect;
-
-chai.use(chaiHttp);
-describe("/GET book", () => {
-	it("it should GET the homepage", (done) => {
-		chai
-			.request(app)
-			.get("/test")
-			.end((err, res) => {
-				res.should.have.status(200);
-				done();
-			});
-	});
-});
